INTRODUCING REAL PONG COLLISION ANGLES
WARNING: Do not merge this code to its mother branch as it modifies the collision logic

The logic for this version of the collision angle mechanics is to consider the paddles behave like half circles.
The next version will introduce the possibility to choose a specific value for the convexity.