Config almost working, a few points to consider:
 - healthcheck is not valid anymore -> trying to find out why
 - the returned url is not working, but a hardcoded url does the job

Which means that we'd better set the default url as a string: https://{SERV_IP}:9000/avatars/{user_id}.jpg
We can test first this url in the front and if it does not return
anything, set the avatar_url to the default case.

Last but not least, there is no purpose to keep the avatar view as it
is.
For an update if the avatar, we can create a service that will
communicate with minio server directly and send the image just after the
user has uploaded it.